Rearrange and refactor entire source. Publicly accessible initialisers are now in the top-level source folder. Other files are organised into folders
