+++ Sources/ASCollectionView/Support/Binding+Sequence.swift
