+++ Sources/ASCollectionView/Support/RandomAccessCollection+Safe.swift
